+++ couchpotato/core/media/movie/providers/torrent/xthor.py
@@ -1,0 +1,11 @@
+from couchpotato.core.logger import CPLog
+from couchpotato.core.media._base.providers.torrent.xthor import Base
+from couchpotato.core.media.movie.providers.base import MovieProvider
+
+log = CPLog(__name__)
+
+autoload = 'xthor'
+
+
+class xthor(MovieProvider, Base):
+    pass
